backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic-settings==2.3.*
python-multipart==0.0.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.29.*
alembic==1.13.*
pgvector==0.3.*
psycopg[binary]==3.2.*
raganything
pypdf==5.9.*
python-docx==1.1.*
openai==1.40.*
tiktoken==0.7.*
boto3==1.34.*
redis[asyncio]==5.0.*
structlog==24.*
httpx==0.27.*
